Created 4 classes, buttons and page's sons.
Added 'Buttons.py' - contains the class of buttons with fields of self, x and y for coordinates around the screen, and an image which will be used as the button itself.

Added 'GameMenuPage.py' - a class of GameMenuPage will represent the objects of the main game menu. Will include buttons at the near future, unfinished.

Added LoginPage - a class of LoginPage will receive a screen definition and will contain the fields of username and password alongwith buttons that will allow to click desired option.

ProfilePage - contains the fields of the necessary information for each user once visiting his profile.